+++ openghg/standardise/_standardise.py
@@ -5,7 +5,7 @@ import warnings
 
 from openghg.objectstore import get_writable_bucket
 from openghg.util import sort_by_filenames
-from openghg.types import optionalPathType, multiPathType
+from openghg.types import pathType, multiPathType
 from numcodecs import Blosc
 import logging
 
@@ -66,9 +66,10 @@ def standardise_surface(
     dataset_source: str | None = None,
     sampling_period: Timedelta | str | None = None,
     calibration_scale: str | None = None,
-    measurement_type: str = "insitu",
+    platform: str | None = None,
+    measurement_type: str | None = None,
     verify_site_code: bool = True,
-    site_filepath: optionalPathType = None,
+    site_filepath: pathType | None = None,
     store: str | None = None,
     update_mismatch: str = "never",
     if_exists: str = "auto",
@@ -105,7 +106,11 @@ def standardise_surface(
         dataset_source: Dataset source name, for example "ICOS", "InGOS", "European ObsPack", "CEDA 2023.06".
         sampling_period: Sampling period as pandas time code, e.g. 1m for 1 minute, 1h for 1 hour
         calibration_scale: Calibration scale for data
-        measurement_type: Type of measurement e.g. insitu, flask
+        platform: Type of measurement platform e.g. "surface-insitu", "surface-flask"
+        measurement_type: Type of measurement. For some source_formats this value is added
+            to the attributes. Platform should be used in preference.
+            If platform is specified and measurement_type is not, this will be
+            set to match the platform.
         verify_site_code: Verify the site code
         site_filepath: Alternative site info file (see openghg/openghg_defs repository for format).
             Otherwise will use the data stored within openghg_defs/data/site_info JSON file by default.
@@ -166,6 +171,7 @@ def standardise_surface(
         dataset_source=dataset_source,
         sampling_period=sampling_period,
         calibration_scale=calibration_scale,
+        platform=platform,
         measurement_type=measurement_type,
         overwrite=overwrite,
         verify_site_code=verify_site_code,