+++ openghg/store/_obssurface.py
@@ -10,7 +10,7 @@ from xarray import Dataset
 from openghg.standardise.meta import align_metadata_attributes
 from openghg.store import DataSchema
 from openghg.store.base import BaseStore
-from openghg.types import multiPathType, pathType, optionalPathType, MetadataAndData, DataOverlapError
+from openghg.types import multiPathType, pathType, MetadataAndData, DataOverlapError
 from collections import defaultdict
 
 logger = logging.getLogger("openghg.store")
@@ -31,7 +31,7 @@ class ObsSurface(BaseStore):
         metadata: dict,
         file_metadata: dict,
         precision_data: bytes | None = None,
-        site_filepath: optionalPathType = None,
+        site_filepath: pathType | None = None,
     ) -> list[dict]:
         """Reads binary data passed in by serverless function.
         The data dictionary should contain sub-dictionaries that contain
@@ -119,9 +119,10 @@ class ObsSurface(BaseStore):
         dataset_source: str | None = None,
         sampling_period: Timedelta | str | None = None,
         calibration_scale: str | None = None,
+        platform: str | None = None,
         measurement_type: str = "insitu",
         verify_site_code: bool = True,
-        site_filepath: optionalPathType = None,
+        site_filepath: pathType | None = None,
         update_mismatch: str = "never",
         if_exists: str = "auto",
         save_current: str = "auto",
@@ -158,7 +159,11 @@ class ObsSurface(BaseStore):
                 before data is finalised.
             dataset_source: Dataset source name, for example "ICOS", "InGOS", "European ObsPack", "CEDA 2023.06"
             sampling_period: Sampling period in pandas style (e.g. 2H for 2 hour period, 2m for 2 minute period).
-            measurement_type: Type of measurement e.g. insitu, flask
+            platform: Type of measurement platform e.g. "surface-insitu", "surface-flask"
+            measurement_type: Type of measurement. For some source_formats this value is added
+                to the attributes. Platform should be used in preference.
+                If platform is specified and measurement_type is not, this will be
+                set to match the platform.
             verify_site_code: Verify the site code
             site_filepath: Alternative site info file (see openghg/openghg_defs repository for format).
                 Otherwise will use the data stored within openghg_defs/data/site_info JSON file by default.
@@ -209,6 +214,7 @@ class ObsSurface(BaseStore):
             clean_string,
             format_inlet,
             format_data_level,
+            format_platform,
             evaluate_sampling_period,
             check_and_set_null_variable,
             hash_file,
@@ -243,15 +249,22 @@ class ObsSurface(BaseStore):
 
         sampling_period = evaluate_sampling_period(sampling_period)
 
+        platform = format_platform(platform, data_type=self._data_type)
+
+        if measurement_type is None and platform is not None:
+            measurement_type = platform
+
         # Ensure we have a clear missing value for data_level, data_sublevel
         data_level = format_data_level(data_level)
         if data_sublevel is not None:
             data_sublevel = str(data_sublevel)
 
+        platform = check_and_set_null_variable(platform)
         data_level = check_and_set_null_variable(data_level)
         data_sublevel = check_and_set_null_variable(data_sublevel)
         dataset_source = check_and_set_null_variable(dataset_source)
 
+        platform = clean_string(platform)
         data_level = clean_string(data_level)
         data_sublevel = clean_string(data_sublevel)
         dataset_source = clean_string(dataset_source)